+++ CHANGELOG.md
@@ -6,6 +6,12 @@ also be mirrored to the Github releases page.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Get operation mode for Timer0
+
 ## [0.5.2] - 2026-04-26
 
 ### Added